requirements.txt
aiogram==3.13
httpx==0.27
aiogoogle==5.11
ydb[yc]==3.17
python-dotenv==1.0